refactor(autostockpile): 拆分识别结果与扫描逻辑 - 抽出节点常量、地区参数解析和 pipeline override 构造 - 分离额度解析、商品扫描与 RecognitionDetail 读取 helper - 收窄 recognition、quantity 和 types 的职责边界
