+++ agent/go-service/autostockpile/nodes.go
@@ -1,0 +1,25 @@
+package autostockpile
+
+const (
+	autoStockpileComponent = "autostockpile"
+
+	autoStockpileSelectItemActionName        = "AutoStockpile.SelectItem"
+	autoStockpileReconcileDecisionActionName = "AutoStockpile.ReconcileDecision"
+	autoStockpileRecognitionName             = "AutoStockpile.Recognition"
+
+	decisionAttachNodeName         = "AutoStockpileDecisionAttach"
+	relayNodeDecisionReadyNodeName = "AutoStockpileRelayNodeDecisionReady"
+	selectedGoodsClickNodeName     = "AutoStockpileSelectedGoodsClick"
+	selectedGoodsClickResetY       = 180
+	swipeMaxNodeName               = "AutoStockpileSwipeMax"
+	swipeSpecificQuantityNodeName  = "AutoStockpileSwipeSpecificQuantity"
+	skipNodeName                   = "AutoStockpileSkip"
+	findMarketMarkNodeName         = "AutoStockpileFindMarketMark"
+	overflowQuotaNodeName          = "AutoStockpileGetQuota"
+	overflowQuotaAdditionNodeName  = "AutoStockpileGetQuotaAddition"
+	locateGoodsNodeName            = "AutoStockpileLocateGoods"
+	goodsPriceNodeName             = "AutoStockpileGetGoods"
+	stockBillNodeName              = "AutoStockpileGetStockBill"
+	autoStockpileDecisionValleyIV  = "AutoStockpileDecisionValleyIV"
+	autoStockpileDecisionWuling    = "AutoStockpileDecisionWuling"
+)
