+++ agent/go-service/autostockpile/register.go
@@ -14,7 +14,7 @@ func Register() {
 			Msg("failed to init item map during registration, OCR name matching may be disabled")
 	}
 
-	maa.AgentServerRegisterCustomAction("AutoStockpile.SelectItem", &SelectItemAction{})
-	maa.AgentServerRegisterCustomAction("AutoStockpile.ReconcileDecision", &ReconcileDecisionAction{})
-	maa.AgentServerRegisterCustomRecognition("AutoStockpile.Recognition", &ItemValueChangeRecognition{})
+	maa.AgentServerRegisterCustomAction(autoStockpileSelectItemActionName, &SelectItemAction{})
+	maa.AgentServerRegisterCustomAction(autoStockpileReconcileDecisionActionName, &ReconcileDecisionAction{})
+	maa.AgentServerRegisterCustomRecognition(autoStockpileRecognitionName, &ItemValueChangeRecognition{})
 }
